+++ buildSrc/src/main/kotlin/com/danrusu/pods4k/immutableArrays/core/ImmutableArrayFactoryGenerator.kt
@@ -81,7 +81,7 @@ private fun FileSpec.Builder.addImmutableArrayOf() {
         ) {
             addGenericTypes(baseType.type)
 
-            statement("return ${baseType.generatedClassName}.copyOf(copy = values, startIndex = 0, size = values.size)")
+            statement("return ${baseType.generatedClassName}.copyFrom(source = values, startIndex = 0, size = values.size)")
         }
     }
 }