+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableBooleanArray.kt
@@ -345,7 +345,7 @@ public value class ImmutableBooleanArray @PublishedApi internal constructor(
         if (n == 0) return EMPTY
         if (n >= size) return this
 
-        return ImmutableBooleanArray.copyOf(copy = values, startIndex = 0, size = n)
+        return ImmutableBooleanArray.copyFrom(source = values, startIndex = 0, size = n)
     }
 
     /**
@@ -370,7 +370,7 @@ public value class ImmutableBooleanArray @PublishedApi internal constructor(
         if (n == 0) return EMPTY
         if (n >= size) return this
 
-        return ImmutableBooleanArray.copyOf(copy = values, startIndex = size - n, size = n)
+        return ImmutableBooleanArray.copyFrom(source = values, startIndex = size - n, size = n)
     }
 
     /**
@@ -498,7 +498,12 @@ public value class ImmutableBooleanArray @PublishedApi internal constructor(
         if (firstIndex == 0) return Pair(EMPTY, this)
         if (firstIndex == size) return Pair(this, EMPTY)
 
-        val first = ImmutableBooleanArray.copyOf(copy = buffer, startIndex = 0, size = firstIndex)
+        val first = ImmutableBooleanArray.copyFrom(
+            source = buffer,
+            startIndex = 0,
+            size =
+            firstIndex,
+        )
         val second = ImmutableBooleanArray(size - first.size) { buffer[size - it - 1] }
         return Pair(first, second)
     }
@@ -584,15 +589,15 @@ public value class ImmutableBooleanArray @PublishedApi internal constructor(
          * Returns an ImmutableBooleanArray with the first [size] elements copied from [copy]
          * starting from [startIndex].
          */
-        public fun copyOf(
-            copy: BooleanArray,
+        public fun copyFrom(
+            source: BooleanArray,
             startIndex: Int,
             size: Int,
         ): ImmutableBooleanArray {
             if (size == 0) return EMPTY
 
             val backingArray = BooleanArray(size)
-            System.arraycopy(copy, startIndex, backingArray, 0, size)
+            System.arraycopy(source, startIndex, backingArray, 0, size)
             return ImmutableBooleanArray(backingArray)
         }
     }
@@ -681,7 +686,7 @@ public value class ImmutableBooleanArray @PublishedApi internal constructor(
                 0 -> return EMPTY
                 values.size -> return ImmutableBooleanArray(values)
             }
-            return copyOf(copy = values, startIndex = 0, size = size)
+            return copyFrom(source = values, startIndex = 0, size = size)
         }
 
         private fun ensureCapacity(minCapacity: Int) {