+++ immutable-arrays/core/src/test/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableArrayTest.kt
@@ -48,31 +48,31 @@ class ImmutableArrayTest {
     fun `copyOf validation`() {
         with(arrayOf("one", "two", "three")) {
             expectThrows<ArrayIndexOutOfBoundsException> {
-                ImmutableArray.copyOf(copy = this, startIndex = -1, size = 1)
+                ImmutableArray.copyFrom(source = this, startIndex = -1, size = 1)
             }
 
             expectThrows<ArrayIndexOutOfBoundsException> {
-                ImmutableArray.copyOf(copy = this, startIndex = 3, size = 1)
+                ImmutableArray.copyFrom(source = this, startIndex = 3, size = 1)
             }
 
             expectThrows<ArrayIndexOutOfBoundsException> {
-                ImmutableArray.copyOf(copy = this, startIndex = 1, size = 3)
+                ImmutableArray.copyFrom(source = this, startIndex = 1, size = 3)
             }
 
             expectThrows<ArrayIndexOutOfBoundsException> {
-                ImmutableArray.copyOf(copy = this, startIndex = 0, size = 4)
+                ImmutableArray.copyFrom(source = this, startIndex = 0, size = 4)
             }
 
-            expectThat(ImmutableArray.copyOf(copy = this, startIndex = 0, size = 0))
+            expectThat(ImmutableArray.copyFrom(source = this, startIndex = 0, size = 0))
                 .hasSize(0)
 
-            expectThat(ImmutableArray.copyOf(copy = this, startIndex = 0, size = 1))
+            expectThat(ImmutableArray.copyFrom(source = this, startIndex = 0, size = 1))
                 .isEqualTo(immutableArrayOf("one"))
 
-            expectThat(ImmutableArray.copyOf(copy = this, startIndex = 1, size = 2))
+            expectThat(ImmutableArray.copyFrom(source = this, startIndex = 1, size = 2))
                 .isEqualTo(immutableArrayOf("two", "three"))
 
-            expectThat(ImmutableArray.copyOf(copy = this, startIndex = 0, size = 3))
+            expectThat(ImmutableArray.copyFrom(source = this, startIndex = 0, size = 3))
                 .isEqualTo(immutableArrayOf("one", "two", "three"))
         }
     }