+++ changelog.md
@@ -9,7 +9,7 @@ _Date TBD_
 
 **New Features:**
 
-* Add `copyOf(regularArray, startIndex, size)` companion factory function
+* Add `copyFrom(regularArray, startIndex, size)` companion factory function
 * Add `randomOrNull(Random)` method
 * Add `randomOrNull()` method
 * Add `random(Random)` method
@@ -17,8 +17,9 @@ _Date TBD_
 
 **Performance Improvements:**
 
-* Use `copyOf` companion factory function in `immutableArrayOf(...)` and `regularArray.toImmutableArray()` to avoid
-  creating a temporary builder.
+* Update `immutableArrayOf(...)` to avoid creating a temporary builder by using the `copyFrom` factory function
+* Update `regularArray.toImmutableArray()` to avoid creating a temporary builder by using the `copyFrom` factory
+  function
 * Update `plus` to concatenate an element with an Immutable Array without using builders.
 * Update `plus` to concatenate 2 Immutable Arrays without using builders.
 * Update `sortedWith` on `ImmutableArray<T>` to copy the elements with `arraycopy`.